增加特效

+++ src/sample-robot/pillbag.py
@@ -1,5 +1,6 @@
-from pygamescratch import Sprite, pygs
+import random
 
+from pygamescratch import Sprite, pygs
 from store import Store
 
 
@@ -7,14 +8,17 @@ class PillBag(Sprite):
 
     def __init__(self, center_x=0, center_y=0):
         super().__init__("pillbag", center_x, center_y)
-        self.point(90)
+        self.point(random.randint(-90, 270))
         self.m = 1
         self.got = False
 
     def action(self):
         self.move(self.m)
         if self.touching_edge():
-            self.delete()
+            if self.got:
+                self.delete()
+            else:
+                self.point(random.randint(-90, 270))
         if not self.got:
             if self.get_touching_sprite("robot"):
                 self.got = True
